queue_store() rtns if msg.dst_ip was in the list.
The specs *say* that ODR should not send duplicate RREQ's. Apparently
this also means that if a app msg has been put in the queue and an RREQ
for that dst IP sent, then another app msg with the same dst IP should
not generate another RREQ.

If somehow all these RREQ's are dropped before reaching a node which can
RREP then these msgs will remain forever in the queue, meaning all
subsequent msgs to this dst IP will not generate RREQs, and never go
through.